+++ src/utils/utils.c
@@ -6,12 +6,12 @@ double getClock() {
     return (double)ts.tv_sec + (double)ts.tv_nsec / BILLION; // seconds with high precision
 }
 
-long long calculateMakespan() {
-    long long totalMakespan = 0;
+int calculateMakespan() {
+    int totalMakespan = 0;
 
     // Calculate total makespan
-    for (long long i = 0; i < totalJobs; i++) {
-        for (long long j = 0; j < totalMachines; j++) {
+    for (int i = 0; i < totalJobs; i++) {
+        for (int j = 0; j < totalMachines; j++) {
             totalMakespan = (totalMakespan > (tasks[i][j].scheduling + tasks[i][j].time))
                 ? totalMakespan
                 : (tasks[i][j].scheduling + tasks[i][j].time);
